Split Verifier into multiple files, more splitting may be necessary
